fix(canvas): align desktop sketchpad contracts
Share canvas code validation, match server field initialization and
fragment defaults, register the flag, and split field utilities.

Generated-By: PostHog Desktop
Task-Id: 617b374c-f922-473b-96a7-f61b557c16ea
