+++ products/desktop/packages/shared/src/feature-flag-keys.json
@@ -32,6 +32,7 @@
   "SAVED_SEARCHES_RAIL_FLAG": "posthog-desktop-saved-searches-rail",
   "SELF_DRIVING_SETUP_TASK_FLAG": "posthog-code-self-driving-setup-task",
   "SIGNALS_PR_REFUNDS_FLAG": "signals-pr-refunds",
+  "SKETCHPADS_FLAG": "posthog-desktop-sketchpads",
   "SPOKEN_NARRATION_FLAG": "posthog-code-spoken-narration",
   "TASK_ANALYSIS_FLAG": "posthog-code-task-analysis",
   "TASK_COST_VISIBLE_FLAG": "posthog-code-task-cost-visible",
