+++ products/desktop/packages/shared/src/flags.ts
@@ -118,4 +118,4 @@ export const CUSTOM_IMAGES_FEATURE_FLAG =
 export const PI_HARNESS_FLAG = featureFlagKeys.PI_HARNESS_FLAG;
 export const TWIG_CLOUD_MODE_FLAG = featureFlagKeys.TWIG_CLOUD_MODE_FLAG;
 export const USER_SPEND_LIMIT_FLAG = featureFlagKeys.USER_SPEND_LIMIT_FLAG;
-export const SKETCHPADS_FLAG = "posthog-desktop-sketchpads";
+export const SKETCHPADS_FLAG = featureFlagKeys.SKETCHPADS_FLAG;
