+++ products/desktop/packages/shared/src/sketchpad/index.ts
@@ -1,8 +1,10 @@
+export * from "./diffText";
 export * from "./egress";
 export * from "./fields";
 export * from "./format";
 export * from "./fragmentCodeGuard";
 export * from "./ops";
+export * from "./orderKey";
 export * from "./paths";
 export * from "./presence";
 export * from "./protocol";
