+++ products/desktop/packages/shared/src/sketchpad/presence.ts
@@ -1,23 +1,14 @@
 import { z } from "zod";
-import { sketchpadViewportSchema } from "./protocol";
 import {
-  SKETCHPAD_FIELD_ID_MAX_CHARS,
-  sketchpadLogEntrySchema,
-} from "./schemas";
+  sketchpadPresenceCaretSchema,
+  sketchpadViewportSchema,
+} from "./protocol";
+import { sketchpadLogEntrySchema, sketchpadUserSchema } from "./schemas";
 
 export const SKETCHPAD_PRESENCE_STALE_MS = 10_000;
 export const SKETCHPAD_PRESENCE_INTERVAL_MS = 100;
 export const SKETCHPAD_PRESENCE_MAX_SELECTED_IDS = 50;
 export const SKETCHPAD_PRESENCE_MAX_CARETS = 4;
-
-export const sketchpadPresenceCaretSchema = z.object({
-  key: z.string().max(128),
-  anchor: z.string().max(SKETCHPAD_FIELD_ID_MAX_CHARS).nullable(),
-  focus: z.string().max(SKETCHPAD_FIELD_ID_MAX_CHARS).nullable(),
-});
-export type SketchpadPresenceCaret = z.infer<
-  typeof sketchpadPresenceCaretSchema
->;
 
 export const sketchpadPresencePointSchema = z.object({
   x: z.number(),
@@ -43,12 +34,9 @@ export type SketchpadPresenceInput = z.infer<
   typeof sketchpadPresenceInputSchema
 >;
 
-export const sketchpadPresenceSchema = sketchpadPresenceInputSchema.extend({
-  userId: z.number().optional(),
-  userUuid: z.string().optional(),
-  userName: z.string().optional(),
-  userEmail: z.string().optional(),
-});
+export const sketchpadPresenceSchema = sketchpadPresenceInputSchema.extend(
+  sketchpadUserSchema.shape,
+);
 export type SketchpadPresence = z.infer<typeof sketchpadPresenceSchema>;
 
 export const sketchpadStreamEventSchema = z.discriminatedUnion("type", [
